Implementacion y correcion del ourvoice
