+++ src/assets/style/index.scss
@@ -6,7 +6,7 @@
 // Components
 @import './components/footer.scss';
 @import './components/navbar.scss';
-@import './components/menu.scss';
+@import './components/ourvoice.scss';
 @import './components/menu.scss';
 * {
   margin: 0px;
