+++ src/components/Sections/Ourvoice/Container.js
@@ -1,0 +1,13 @@
+import ContainerLeft from './ContainerLeft';
+import ContainerRight from './ContainerRight';
+
+function Container() {
+  return (
+    <div className="ourvoice__container">
+      <ContainerLeft />
+      <ContainerRight />
+    </div>
+  );
+}
+
+export default Container;
