+++ src/components/Sections/Ourvoice/ContainerLeft.js
@@ -1,0 +1,10 @@
+function ContainerLeft() {
+  return (
+    <div className="ourvoice__container-left">
+      <h2 className="ourvoice__container-left-title">Our voice</h2>
+      <h1 className="ourvoice__container-left-post">Post</h1>
+    </div>
+  );
+}
+
+export default ContainerLeft;
