+++ src/components/Sections/Ourvoice/ContainerRight.js
@@ -1,0 +1,69 @@
+function ContainerRight() {
+  return (
+    <div className="ourvoice__container-right">
+      <div className="ourvoice__container-right-item">
+        {/*----------------ITEM 1-----------*/}
+
+        <h1 className="ourvoice__container-right-item-title">
+          Let's talk about react
+        </h1>
+        <p className="ourvoice__container-right-item-content">
+          A short tour of react to a technology that changed my life and
+          continues to change the industry.
+        </p>
+        <div className="ourvoice__container-right-item-information">
+          <p className="ourvoice__container-right-item-information-readers">
+            300 readers
+          </p>
+          <p className="ourvoice__container-right-item-information-links">
+            Read a little more
+          </p>
+        </div>
+      </div>
+
+      {/*----------------ITEM 2-----------*/}
+
+      <div className="ourvoice__container-right-item">
+        <h1 className="ourvoice__container-right-item-title">
+          Elegan code in react
+        </h1>
+        <p className="ourvoice__container-right-item-content">
+          There are many ways to write code in react, but do you know how to do
+          it elegantly?
+        </p>
+
+        <div className="ourvoice__container-right-item-information">
+          <p className="ourvoice__container-right-item-information-readers">
+            300 readers
+          </p>
+          <p className="ourvoice__container-right-item-information-links">
+            Read a little more
+          </p>
+        </div>
+      </div>
+
+      {/*----------------ITEM 3-----------*/}
+
+      <div className="ourvoice__container-right-item">
+        <h1 className="ourvoice__container-right-item-title">
+          Tips from a junior.
+        </h1>
+        <p className="ourvoice__container-right-item-content">
+          Some tips from a junior frontend developer that you may not have known
+          and probably need.
+        </p>
+
+        <div className="ourvoice__container-right-item-information">
+          <p className="ourvoice__container-right-item-information-readers">
+            300 readers
+          </p>
+          <p className="ourvoice__container-right-item-information-links">
+            Read a little more
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default ContainerRight;
