+++ src/components/Sections/Ourvoice/Ourvoice.js
@@ -1,0 +1,15 @@
+import Container from './Container';
+
+function Ourvoice() {
+  return (
+    <div className="ourvoice">
+      <div className="ourvoice__title">
+        <h2 className="ourvoice__title-hero">Our voice</h2>
+        <h1 className="ourvoice__title-post">Post</h1>
+      </div>
+      <Container />
+    </div>
+  );
+}
+
+export default Ourvoice;
